move NullTime to nulltime.go to allow to use build tags
